requirements.txt
pandas==2.3.2
psycopg2==2.9.10
python-dotenv==1.1.1